R/eport

+++ Application/DTOs/ReportDTO.cs
@@ -13,7 +13,7 @@ namespace Application.DTOs
         public string? ReportToUrl { get; set; }
         public int? ProjectId { get; set; }
 
-        public int? BidId { get; set; }
+        public long? BidId { get; set; }
         public int? CreatedBy { get; set; }
         public int ReportCategoryId { get; set; }
         public string Description { get; set; }
